refactor: any 타입 정리
