+++ src/components/Modals/ModalInput.tsx
@@ -2,11 +2,7 @@ import React from 'react';
 import './Modal.css';
 
 function ModalInput() {
-  return (
-    <>
-      <input className="modalInput" placeholder="내용 입력"></input>
-    </>
-  );
+  return <input className="modalInput" placeholder="내용 입력" />;
 }
 
 export default ModalInput;
