+++ src/components/pages/folder.tsx
@@ -1,16 +1,14 @@
 import Header from '../Header/Header';
 import Footer from '../Footer/Footer';
 import Section from '../Section/Section';
-import ModalPortal from '../../utils/Portal';
+import ModalPortal from '../Portal';
 
 export default function Folder() {
   return (
-    <>
-      <ModalPortal>
-        <Header style="headerFolder" />
-        <Section style="sectionFolder" />
-        <Footer />
-      </ModalPortal>
-    </>
+    <ModalPortal>
+      <Header style="headerFolder" />
+      <Section style="sectionFolder" />
+      <Footer />
+    </ModalPortal>
   );
 }
